Move sanitize_content_filename to network.download
